BLADE-533 move util classes into cli
