+++ cli/src/main/java/com/liferay/blade/cli/util/NodeUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.liferay.project.templates.js.widget.internal;
+package com.liferay.blade.cli.util;
 
 import com.liferay.blade.cli.util.BladeUtil;
 import com.liferay.blade.cli.util.FileUtil;
